02.Serverless-Automation-with-GCP-and-Shelly-Devices/resources/handle-shelly-button/requirements.txt
functions-framework==3.*
cloudevents==1.9.0
google-cloud-pubsub